🦺  reorganização das pastas
Reorganizei as pastas e excluí o que foi pedido.
