+++ frontend/src/views/ApisView.vue
@@ -1,0 +1,17 @@
+
+
+<template>
+    <ListarApi />
+</template>
+
+
+<script>
+    import ListarApi from "@/components/ListarApi.vue";
+    export default {
+      name: 'ListarApiView',
+      components: {
+        ListarApi
+      }
+    }
+</script>
+
